Updated usage of "not expr1 in expr2" to support Elixir 1.5+
Elixir 1.5+ shall be deprecating this operation.
I just added parentheses to fix this and also support earlier Elixir versions.

See the reported warning below:
"not expr1 in expr2" is deprecated. Instead use "expr1 not in expr2" if you require Elixir v1.5+, or "not(expr1 in expr2)" if you have to support earlier Elixir versions.